# perf: vectorize yield uncertainty calculation