@alexander-held 2022-02-09
Here is an example comparing three approaches: a `numpy` solution using matrix multiplications, the old loop-based approach, and the new approach supporting `awkward` arrays with additional depth:
```python
import awkward as ak
import numpy as np


def calc(op):
    print(f"{op} has shape {eval(op).shape}:")
    print(f"{eval(op)}\n")


# setup: 3 variations, 5 bin yields
v = np.asarray([[1, 2, 3, 4, 5], [3, 4, 5, 6, 7], [5, 6, 7, 8, 9]])
M = np.asarray([[1, 1.4, -0.3], [1.4, 1, 1.2], [-0.3, 1.2, 1]])


calc("v")
calc("v.T")
calc("M")
calc("v.T @ M @ v")
calc("np.diag(v.T @ M @ v)")

# loop-based calculation
res = 0
for i in range(3):
    for j in range(3):
        res += v[i] * v[j] * M[i][j]
print(f"loop-based calculation: {res}\n")

# variation with awkward and extra depth in v
# elements of v are now split e.g. by channel
v = ak.from_iter([[[1, 2, 3], [4, 5]], [[3, 4, 5], [6, 7]], [[5, 6, 7], [8, 9]]])
R = M[..., np.newaxis, np.newaxis] * v[np.newaxis, ...]
L = v[:, np.newaxis, ...] * R
print(f"awkward version with extra depth: {np.sum(ak.flatten(L, axis=1), axis=0)}")
```
output:
```
v has shape (3, 5):
[[1 2 3 4 5]
 [3 4 5 6 7]
 [5 6 7 8 9]]

v.T has shape (5, 3):
[[1 3 5]
 [2 4 6]
 [3 5 7]
 [4 6 8]
 [5 7 9]]

M has shape (3, 3):
[[ 1.   1.4 -0.3]
 [ 1.4  1.   1.2]
 [-0.3  1.2  1. ]]

v.T @ M @ v has shape (5, 5):
[[ 76.4  98.8 121.2 143.6 166. ]
 [ 98.8 128.8 158.8 188.8 218.8]
 [121.2 158.8 196.4 234.  271.6]
 [143.6 188.8 234.  279.2 324.4]
 [166.  218.8 271.6 324.4 377.2]]

np.diag(v.T @ M @ v) has shape (5,):
[ 76.4 128.8 196.4 279.2 377.2]

loop-based calculation: [ 76.4 128.8 196.4 279.2 377.2]

awkward version with extra depth: [[76.4, 129, 196], [279, 377]]
```
The `awkward` calculation matches the other two, there are just less digits printed.