@agoose77 2022-02-10
@alexander-held maybe it's worth adding an assert with a message to catch the case where/if the assumption fails?