@alexander-held 2022-02-11
@agoose77 by "the assumption", do you mean that we only sum over an array with three or less dimensions (after flattening the original 4d array)? I cannot think of a way in which this would break, but it also does not hurt to have that in. I'll add an assert, and we can refactor and remove it in the future.