@alexander-held 2022-02-11
I added the assert, #326 will track the task of changing this after an upstream fix.

Thanks to you both for the help with this PR!